Entrega Final

+++ js/shopCart.js
@@ -1,6 +1,5 @@
 let productsInCart = localStorage.getItem("products-in-cart");
-  productsInCart = JSON.parse(productsInCart);
-
+productsInCart = JSON.parse(productsInCart);
 
 const containerCartEmpty = document.querySelector("#cart-empty");
 const containerCartProducts = document.querySelector("#cart-products");
@@ -13,8 +12,9 @@ const buttomBuy = document.querySelector("#cart-actions-buy");
 
 function loadCartProducts() {
 
-    if (productsInCart && productsInCart.lenght > 0) {
-      
+    if (productsInCart && productsInCart.length > 0) {
+
+
         containerCartEmpty.classList.add("disabled");
         containerCartProducts.classList.remove("disabled");
         containerCartActions.classList.remove("disabled");
@@ -44,7 +44,10 @@ function loadCartProducts() {
             <small>Subtotal</small>
             <p>$${product.price * product.amount}</p>
         </div>
-        <button class="cart-product-eliminate" id="${product.id}"><i class="bi bi-trash-fill"></i></button>
+        <button class="cart-product-eliminate" id="${product.id}"> <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash3" viewBox="0 0 16 16">
+        <path d="M6.5 1h3a.5.5 0 0 1 .5.5v1H6v-1a.5.5 0 0 1 .5-.5ZM11 2.5v-1A1.5 1.5 0 0 0 9.5 0h-3A1.5 1.5 0 0 0 5 1.5v1H2.506a.58.58 0 0 0-.01 0H1.5a.5.5 0 0 0 0 1h.538l.853 10.66A2 2 0 0 0 4.885 16h6.23a2 2 0 0 0 1.994-1.84l.853-10.66h.538a.5.5 0 0 0 0-1h-.995a.59.59 0 0 0-.01 0H11Zm1.958 1-.846 10.58a1 1 0 0 1-.997.92h-6.23a1 1 0 0 1-.997-.92L3.042 3.5h9.916Zm-7.487 1a.5.5 0 0 1 .528.47l.5 8.5a.5.5 0 0 1-.998.06L5 5.03a.5.5 0 0 1 .47-.53Zm5.058 0a.5.5 0 0 1 .47.53l-.5 8.5a.5.5 0 1 1-.998-.06l.5-8.5a.5.5 0 0 1 .528-.47ZM8 4.5a.5.5 0 0 1 .5.5v8.5a.5.5 0 0 1-1 0V5a.5.5 0 0 1 .5-.5Z"/>
+      </svg></button>
+        <div class="separator"></div>
     `;
 
             containerCartProducts.append(div);
@@ -72,21 +75,36 @@ function resetBtnEliminate() {
     })
 }
 
+
 function eliminateProduct(e) {
     const idButton = e.currentTarget.id;
     const index = productsInCart.findIndex(product => product.id === idButton);
-    productsInCart.splice(index, 1)
-    loadCartProducts();
-
-    localStorage.setItem("products-in-cart", JSON.stringify(productsInCart));
+    if (index !== -1) {
+        productsInCart.splice(index, 1);
+        loadCartProducts();
+        localStorage.setItem("products-in-cart", JSON.stringify(productsInCart));
+    }
 }
 
 buttomCartEliminate.addEventListener("click", deleteCart);
 
 function deleteCart() {
-    productsInCart.lenght = 0;
-    localStorage.setItem("products-in-cart", JSON.stringify(productsInCart));
-    loadCartProducts()
+
+    Swal.fire({
+        title: '¿Estás seguro?',
+        icon: 'question',
+        html: `Se van a borrar ${productsInCart.reduce((acc, product) => acc + product.amount, 0)} productos.`,
+        showCancelButton: true,
+        focusConfirm: false,
+        confirmButtonText: 'Sí',
+        cancelButtonText: 'No'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            productsInCart.length = 0;
+            localStorage.setItem("products-in-cart", JSON.stringify(productsInCart));
+            loadCartProducts();
+        }
+    })
 }
 
 function loadTotal() {
@@ -97,11 +115,14 @@ function loadTotal() {
 buttomBuy.addEventListener("click", buyCart);
 
 function buyCart() {
-    productsInCart.lenght = 0;
-    localStorage.setItem("products-in-cart", JSON.stringify(productsInCart));
+    productsInCart.length = 0;
 
+    localStorage.setItem("products-in-cart", JSON.stringify(productsInCart));
     containerCartEmpty.classList.add("disabled");
     containerCartProducts.classList.add("disabled");
     containerCartActions.classList.add("disabled");
     containerCartBought.classList.remove("disabled");
+
+
 }
+
